Refactor: Extract OAuth logic to separate module
