+++ README.md
@@ -25,7 +25,7 @@
 ## ✔️ 설치 방법
 ### 필요한 의존성 패키지
 ```
-pip install flask pymongo dnspython jwt certifi requests
+pip install flask pymongo dnspython jwt certifi requests Flask-OAuthlib dotenv
 ```
 
 ## ✔️ 구성원 및 맡은 기능
